[chore] #1 - 프로젝트 구조 세팅
