+++ Tuist/ProjectDescriptionHelpers/InfoPlist.swift
@@ -1,0 +1,11 @@
+import ProjectDescription
+
+extension Project {
+    static let infoPlist: [String: InfoPlist.Value] = [
+        "CFBundleShortVersionString": "1.0",
+        "CFBundleVersion": "1",
+        "UIMainStoryboardFile": "",
+        "UILaunchStoryboardName": "LaunchScreen",
+        "UIUserInterfaceStyle": "Light"
+    ]
+}
