+++ Tuist/ProjectDescriptionHelpers/Project+Templates.swift
@@ -1,32 +1,20 @@
 import ProjectDescription
 
-public enum Environment {
-    public static let bundleId: String = "com.PIMO"
-    public static let organizationName: String = "PIMO"
-    public static let deploymentTarget: DeploymentTarget = .iOS(targetVersion: "16.0", devices: [.iphone])
-}
-
 extension Project {
     public static func app(name: String,
                            platform: Platform,
                            dependencies: [TargetDependency] = [])
     -> Project {
         
-        let infoPlist: [String: InfoPlist.Value] = [
-            "CFBundleShortVersionString": "1.0",
-            "CFBundleVersion": "1",
-            "UIMainStoryboardFile": "",
-            "UILaunchStoryboardName": "LaunchScreen"
-        ]
-        
         let mainTarget = Target(
             name: name,
             platform: platform,
             product: .app,
             bundleId: "\(Environment.bundleId)",
-            infoPlist: .extendingDefault(with: infoPlist),
+            infoPlist: .extendingDefault(with: Project.infoPlist),
             sources: ["\(name)/Sources/**"],
             resources: ["\(name)/Resources/**"],
+            scripts: [.SwiftLintShell],
             dependencies: dependencies
         )
